improvement: add Marshall method for HeuristicStatus and unit tests
