+++ parsers/nmparser/parse_lst_file_test.go
@@ -468,57 +468,3 @@ func TestSetCorrelationsOk(t *testing.T) {
 		assert.Equal(t, tt.correlationsOk, correlationsOk, "Fail :"+tt.context)
 	}
 }
-
-func TestUnmarshalJSON(t *testing.T) {
-	var tests = []struct {
-		value    []byte
-		expected HeuristicStatus
-		context  string
-	}{
-		{
-			value:    []byte(string(`"HeuristicUndefined"`)),
-			expected: HeuristicUndefined,
-			context:  "HeuristicUndefined",
-		},
-		{
-			value:    []byte(string(`"HeuristicTrue"`)),
-			expected: HeuristicTrue,
-			context:  "HeuristicTrue",
-		},
-		{
-			value:    []byte(string(`"HeuristicFalse"`)),
-			expected: HeuristicFalse,
-			context:  "HeuristicFalse",
-		},
-	}
-
-	for _, tt := range tests {
-
-		hs := new(HeuristicStatus)
-		err := hs.UnmarshalJSON(tt.value)
-		assert.Equal(t, nil, err, "Fail :"+tt.context)
-		assert.Equal(t, tt.expected, *hs, "Fail :"+tt.context)
-	}
-}
-
-// func TestMarshalJSON(t *testing.T) {
-// 	var tests = []struct {
-// 		value    HeuristicStatus
-// 		expected string
-// 		context  string
-// 	}{
-// 		{
-// 			value:    HeuristicUndefined,
-// 			expected: "",
-// 			context:  "HeuristicUndefined",
-// 		},
-// 	}
-
-// 	for _, tt := range tests {
-
-// 		hs := new(HeuristicStatus)
-// 		err := hs.MarshalJSON(tt.value)
-// 		assert.Equal(t, nil, err, "Fail :"+tt.context)
-// 		assert.Equal(t, tt.expected, *hs, "Fail :"+tt.context)
-// 	}
-// }
